--- company/templates/company-office-address.html
@@ -1,5 +1,0 @@
-<div id="ed-company-edit-address">
-	<p class="ed-company-edit-address-name">{{ company_name }}</p>
-	<p class="ed-company-edit-address-number">Company number <strong>{{ company_number }}</strong></p>
-	<p>{{ company_address }}</p>
-</div>
