--- company/templates/company-profile-form-classification.html
@@ -1,15 +1,0 @@
-{% extends "form-wrapper.html" %}
-
-{% load static from staticfiles %}
-
-{% block hero_title %}Create your company profile{% endblock %}
-{% block hero_text %}Add your company details to help others find you.{% endblock %}
-{% block form_title %}Your company sector{% endblock%}
-{% block button_next_label %}Next{% endblock%}
-{% block button_finish_label %}Save{% endblock%}
-{% block formarea_container_class %}column-full ed-form-classification-container{% endblock %}
-
-{% block body_js %}
-	{{ block.super }}
-    <script type="text/javascript" src="{% static 'javascripts/company-edit.js' %}"></script>
-{% endblock %}
