--- company/templates/company-profile-form-contact.html
@@ -1,13 +1,0 @@
-{% extends "form-wrapper.html" %}
-{% block head_title %}Nominate your company contact - great.gov.uk{% endblock %}
-{% block form_title %}Nominate a business contact{% endblock%}
-
-{% block form_subtitle %}
-<p>
-	Add your details or nominate someone else in your company to receive enquiries from overseas buyers.
-</p>
-<br>
-{% endblock %}
-
-{% block button_next_label %}Next{% endblock%}
-{% block button_finish_label %}Save{% endblock%}
