--- company/templates/company-profile-form.html
@@ -1,18 +1,0 @@
-{% extends "form-wrapper.html" %}
-
-{% block hero_title %}Build and improve your company profile{% endblock %}
-{% block hero_text %}Add your company details to help others find you.{% endblock %}
-{% block form_title %}Create your business profile{% endblock%}
-{% block button_next_label %}Next{% endblock%}
-{% block button_finish_label %}Save{% endblock%}
-{% block form_subtitle %}
-<p>
-	Enter your company name and contact details, then select the industry or industries you work in.
-</p>
-<p>
-	Add extra information and your company logo to complete your profile. You can come back and edit your profile at any time.
-</p>
-
-<br>
-<h2 class="heading-large push--bottom">Your company details</h2>
-{% endblock form_subtitle %}
