--- company/templates/company-profile-logo-form.html
@@ -1,8 +1,0 @@
-{% extends "form-wrapper.html" %}
-
-{% block form_title %}Upload your company's logo{% endblock%}
-{% block button_next_label %}Next{% endblock%}
-{% block button_finish_label %}Upload file{% endblock%}
-{% block extra_buttons %}
-    <a href="{% url 'company-detail' %}" class="register-btn button" style="color: #4A4A4F; background: #fafafa;" target="_self">Cancel</a>
-{% endblock %}
