--- company/templates/image-too-large.html
@@ -1,21 +1,0 @@
-{% extends 'enrolment-base.html' %}
-{% block sub_header %}{# prevent the subheader from showing on form pages #}{% endblock %}
-
-{% block content %}
-    <section>
-        <div class="outer-container">
-            <div class="grid-row">
-                <span class="column-full ed-hero-sub-header-title">
-                    <h1 class="heading-xlarge">Your image was too large</h1>
-                </span>
-            </div>
-            <br>
-            <p>
-                <a href="{{ request.META.HTTP_REFERER }}" target="_self">Click here</a> to go back and select a smaller image. Images can be no more than 5 Megabytes.
-            </p>
-            <br>
-            <div class="clearfix"></div>
-        </div>
-    </section>
-
-{% endblock %}
