--- company/templates/supplier-case-study-basic-form.html
@@ -1,11 +1,0 @@
-{% extends "form-wrapper.html" %}
-{% block form_title %}Create case study or project{% endblock %}
-{% block form_subtitle %}
-<p>
-	Describe something significant your company has done that will be relevant to international buyers. For example, you could add details of a significant collaboration on a recent international project or a major export deal.
-</p>
-<br>
-{% endblock %}
-{% block extra_buttons %}
-    <a href="{% url 'company-detail' %}" class="register-btn button" style="color: #4A4A4F; background: #fafafa;" target="_self">Cancel</a>
-{% endblock %}
