--- company/templates/supplier-case-study-error.html
@@ -1,4 +1,0 @@
-{% extends 'enrolment-base.html' %}
-
-{% block hero_title %}There was a problem updating your case study.{% endblock %}
-{% block hero_text %}Please try again.{% endblock %}
