--- company/templates/supplier-case-study-rich-media-form.html
@@ -1,3 +1,0 @@
-{% extends "form-wrapper.html" %}
-{% block form_title %}Add images and testimonial{% endblock %}
-{% block button_finish_label %}Save{% endblock%}
