--- company/tests/helpers.py
@@ -1,13 +1,0 @@
-from io import BytesIO
-
-from PIL import Image, ImageDraw
-
-
-def create_test_image(extension):
-    image = Image.new("RGB", (300, 50))
-    draw = ImageDraw.Draw(image)
-    draw.text((0, 0), "This text is drawn on image")
-    byte_io = BytesIO()
-    image.save(byte_io, extension)
-    byte_io.seek(0)
-    return byte_io
