--- core/templates/robots.txt
@@ -1,5 +1,0 @@
-{% extends 'robots.txt' %}
-{% block records %}
-{{ block.super }}
-Disallow: {% url 'request-payload-too-large' %}
-{% endblock %}
