+++ enrolment/helpers.py
@@ -1,60 +1,20 @@
 from collections import OrderedDict
 from functools import partial
-import http
 import logging
 import urllib
 from urllib.parse import urljoin
 
-from directory_ch_client.company import CompanyCHClient
 from django.conf import settings
-from django.forms import ValidationError
-from django.shortcuts import render
 
 import requests
-from requests.exceptions import RequestException
 
 from directory_api_client.client import api_client
-from enrolment import validators
 
 
-COMPANIES_HOUSE_PROFILE_SESSION_KEY = 'ch_profile'
 MESSAGE_AUTH_FAILED = 'Auth failed with Companies House'
-MESSAGE_NETWORK_ERROR = 'A network error occurred'
-COMPANIES_HOUSE_DATE_FORMAT = '%Y-%m-%d'
 
 
 logger = logging.getLogger(__name__)
-
-
-def get_company_from_companies_house(company_number):
-    response = CompaniesHouseClient.retrieve_profile(number=company_number)
-    response.raise_for_status()
-    return response.json()
-
-
-def store_companies_house_profile_in_session(session, company_number):
-    details = get_company_from_companies_house(company_number)
-    session[COMPANIES_HOUSE_PROFILE_SESSION_KEY] = {
-        'company_name': details['company_name'],
-        'company_status': details['company_status'],
-        'date_of_creation': details.get('date_of_creation'),
-        'company_number': company_number,
-        'registered_office_address': details['registered_office_address']
-    }
-    session.modified = True
-
-
-def halt_validation_on_failure(*all_validators):
-    """
-    Django runs all validators on a field and shows all errors. Sometimes this
-    is undesirable: we may want the validators to stop on the first error.
-
-    """
-    def inner(value):
-        for validator in all_validators:
-            validator(value)
-    inner.inner_validators = all_validators
-    return [inner]
 
 
 def has_company(sso_session_id):
@@ -84,12 +44,6 @@ class CompaniesHouseClient:
         'oauth2-token': make_oauth2_url('oauth2/token'),
     }
     session = requests.Session()
-    companies_house_client = CompanyCHClient(
-        base_url=settings.DIRECTORY_CH_SEARCH_CLIENT_BASE_URL,
-        api_key=settings.DIRECTORY_CH_SEARCH_CLIENT_API_KEY,
-        sender_id=settings.DIRECTORY_CH_SEARCH_CLIENT_SENDER_ID,
-        timeout=settings.DIRECTORY_CH_SEARCH_CLIENT_DEFAULT_TIMEOUT,
-    )
 
     @classmethod
     def get_auth(cls):
@@ -98,39 +52,9 @@ class CompaniesHouseClient:
     @classmethod
     def get(cls, url, params={}):
         response = cls.session.get(url=url, params=params, auth=cls.get_auth())
-        if response.status_code == http.client.UNAUTHORIZED:
+        if response.status_code == 403:
             logger.error(MESSAGE_AUTH_FAILED)
         return response
-
-    @classmethod
-    def retrieve_profile(cls, number):
-        if settings.FEATURE_FLAGS['INTERNAL_CH_ON']:
-            return cls.companies_house_client.get_company_profile(
-                company_number=number
-            )
-        else:
-            url = cls.endpoints['profile'].format(number=number)
-            return cls.get(url)
-
-    @classmethod
-    def retrieve_address(cls, number):
-        if settings.FEATURE_FLAGS['INTERNAL_CH_ON']:
-            return cls.companies_house_client.get_company_registered_address(
-                company_number=number
-            )
-        else:
-            url = cls.endpoints['address'].format(number=number)
-            return cls.get(url)
-
-    @classmethod
-    def search(cls, term):
-        if settings.FEATURE_FLAGS['INTERNAL_CH_ON']:
-            return cls.companies_house_client.search_companies(
-                query=term
-            )
-        else:
-            url = cls.endpoints['search']
-            return cls.get(url, params={'q': term})
 
     @classmethod
     def make_oauth2_url(cls, redirect_uri, company_number):
@@ -154,57 +78,3 @@ class CompaniesHouseClient:
             ('redirect_uri', redirect_uri),
         ])
         return cls.session.post(url=url + '?' + urllib.parse.urlencode(params))
-
-
-def get_date_of_creation_from_session(session):
-    return session[
-        COMPANIES_HOUSE_PROFILE_SESSION_KEY
-    ]['date_of_creation']
-
-
-def get_company_number_from_session(session):
-    key = 'company_number'
-    return session.get(COMPANIES_HOUSE_PROFILE_SESSION_KEY, {}).get(key)
-
-
-def get_company_name_from_session(session):
-    return get_company_from_session(session)['company_name']
-
-
-def get_company_status_from_session(session):
-    return get_company_from_session(session)['company_status']
-
-
-def get_company_address_from_session(session):
-    return get_company_from_session(session)['registered_office_address']
-
-
-def get_company_from_session(session):
-    return session.get(COMPANIES_HOUSE_PROFILE_SESSION_KEY, {})
-
-
-def store_companies_house_profile_in_session_and_validate(
-        session, company_number
-):
-    try:
-        store_companies_house_profile_in_session(
-            session=session,
-            company_number=company_number,
-        )
-    except RequestException as error:
-        error = getattr(error, 'response', None)
-        status_code = getattr(error, 'status_code', None)
-        if status_code == http.client.NOT_FOUND:
-            raise ValidationError(validators.MESSAGE_COMPANY_NOT_FOUND)
-        else:
-            raise ValidationError(validators.MESSAGE_COMPANY_ERROR)
-    else:
-        company_status = get_company_status_from_session(
-            session
-        )
-        validators.company_active(company_status)
-        validators.company_unique(company_number)
-
-
-def get_error_response(request, context):
-    return render(request, 'enrolment-error.html', context)
