--- enrolment/templates/aims-form.html
@@ -1,2 +1,0 @@
-{% extends "form-wrapper.html" %}
-{% block form_title %}Your exporting aims{% endblock%}
