--- enrolment/templates/company-form.html
@@ -1,39 +1,0 @@
-{% extends "form-wrapper.html" %}
-{% load static from staticfiles %}
-
-{% block button_finish_label %}{{ sso_is_logged_in|yesno:"Continue,Create account" }}{% endblock%}
-
-{% block progress_indicator %}{% endblock %}
-
-{% block form_container_title %}
-	<div class="outer-container">
-		<h2 class="heading-large">Confirm your company</h2>
-	</div>
-{% endblock %}
-
-{% block form_subtitle %}
-<div class="grid-row">
-	<div class="column-two-thirds">
-		<div class="input-field-container" id="ed-company-edit-address-container">
-			<dl>
-				<dt>Registered name</dt>
-				<dd>{{ company_name }}</dd>
-				<dt>Company number</dt>
-				<dd>{{ company_number }}</dd>
-				<dt>Registered address</dt>
-				<dd>{{ company_address }}</dd>
-			</dl>
-			<div class="{{ form.confirmed.css_classes }} authorised-checkbox">
-				{{ form.confirmed.errors }}
-				{{ form.confirmed }}
-			</div>
-		</div>
-		{% if not sso_is_logged_in %}
-			<div id="ed-account-create-explainer">
-				<p>To create a business profile for this company you must <a href="{{ sso_login_url }}" class="link">Sign in</a> to your account.</p>
-				<p>If you do not have a great.gov.uk account you can create one here.</p>
-			</div>
-		{% endif %}
-	</div>
-</div>
-{% endblock %}
