--- enrolment/templates/landing-page-old.html
@@ -1,72 +1,0 @@
-{% extends 'base.html' %}
-{% load staticfiles %}
-{% load trans from i18n %}
-{% block head_css %}
-	{{ block.super }}
-	<link href="{% static 'landing-page.css' %}" media="all" rel="stylesheet" />
-{% endblock %}
-{% block sub_header %}{# prevent the subheader from showing on form pages #}{% endblock %}
-{% block hero_title %}
-Get seen by motivated overseas buyers
-{% endblock %}
-
-{% block css_layout_class %}
-	{{ block.super }}
-	fabhome fabhome-old
-{% endblock %}
-
-{% block content %}
-	<section id="fabhome-intro" class="outer-container">
-		<h1 class="element-within company-name-title">
-			<span>{% trans "Get promoted internationally" %}</span>
-			<span class="sub">{% trans "with a great.gov.uk business profile" %}</span>
-		</h1>
-		<div>
-			{% include 'register-company-number-form.html' %}
-		</div>
-
-		<p class="message">{% trans "Connect directly with international buyers" %}</p>
-	</section>
-
-	<section id="fabhome-benefits">
-		<div class="outer-container">
-			<h2 class="visuallyhidden">{% trans "Benefit of a business profile" %}</h2>
-			<p>{% trans "Your free great.gov.uk business profile will:" %}</p>
-			<ul>
-				<li>{% trans "let international buyers get in touch with your company's sales team" %}</li>
-				<li>{% trans "showcase your company's outstanding projects and experience to give buyers insight into what you do" %}</li>
-				<li>{% trans "give your company international credibility by displaying independent data from Companies House" %}</li>
-				<li>{% trans "give companies looking to buy from your industry an easy way to find you" %}</li>
-			</ul>
-			<img id="fabhome-screenshot" src="{% static 'images/960-FAS-preview-case-study-435px.png' %}" alt="" />
-		</div>
-	</section>
-
-	<section id="fabhome-howto">
-		<div class="outer-container">
-			<h2>{% trans "Get your profile in 3 simple steps" %}</h2>
-			<ol>
-				<li class="step1">
-					<strong aria-describedby="step1desc">{% trans "Sign up" %}</strong>
-					<span id="step1desc">{% trans "Enter your Companies House number to start building your business profile." %}</span>
-				</li>
-
-				<li class="step2">
-					<strong aria-describedby="step2desc">{% trans "Verify" %}</strong>
-					<span id="step2desc">{% trans "Prove your affiliation to your company." %}</span>
-				</li>
-
-				<li class="step3">
-					<strong aria-describedby="step3desc">{% trans "Promote" %}</strong>
-					<span id="step3desc">{% trans "Describe what your company does and include examples of your best work." %}</span>
-				</li>
-			</ol>
-		</div>
-
-		<div id="fabhome-howto-lookup-area" {% if user_has_company %}style="visibility: hidden;"{% endif %}>
-			<div class="outer-container">
-				{% include 'register-company-number-form.html' %}
-			</div>
-		</div>
-	</section>
-{% endblock content %}
