--- enrolment/templates/register-company-number-form.html
@@ -1,27 +1,0 @@
-{% load trans from i18n %}
-
-{# company number lookup. shows only when js is turned off#}
-<form class="js-disabled-only register-company-number-form" method="post" action="{% url 'index' %}">
-	<div class="errors">
-		{{ form.company_number.errors }}
-	</div>
-	<label>
-		<span class="label">{% trans "Enter your Companies House number" %}</span>
-		<input name="company_number"type="text" placeholder="{% trans 'Companies House number' %}" />
-	</label>
-	<button class="button button-cta" type="submit">{% trans 'Get promoted' %}</button>
-</form>
-
-{# company name lookup. shows only when js is turned on #}
-<form class="js-enabled-only register-company-number-form" method="post" action="{% url 'index' %}">
-	<div class="errors">
-		{{ form.company_number.errors }}
-	</div>
-	<label>
-		<span class="label">{% trans "Enter your company name" %}</span>
-		<input name="company_name" type="text" placeholder="{% trans 'Company name' %}" autofocus="autofocus" />
-	</label>
-	<input type="hidden" name="company_number" />
-	<button class="button button-cta" type="submit">{% trans 'Get promoted' %}</button>
-</form>
-
