+++ requirements.in
@@ -9,13 +9,12 @@ django-redis==4.8.0
 gunicorn==19.5.0
 raven==5.19.0
 whitenoise==3.1
-phonenumberslite==7.7.4
-pillow==4.2.1
 directory_validators==5.0.0
 directory-healthcheck==1.0.1
-directory-components==8.3.1
-directory-ch-client==1.0.0
+directory-components==15.1.1
 directory-sso-api-client==4.1.1
 directory-api-client==13.0.0
 requests[security]==2.21.0
 urllib3>=1.24.2,<2.0.0
+directory-constants==16.7.2
+djangorestframework==3.9.1
