+++ requirements.txt
@@ -5,6 +5,7 @@
 #    pip-compile --output-file requirements.txt requirements.in
 #
 asn1crypto==0.24.0        # via cryptography
+attrs==19.1.0             # via jsonschema
 beautifulsoup4==4.6.3     # via directory-components
 certifi==2018.4.16        # via requests
 cffi==1.11.5              # via cryptography
@@ -12,39 +13,38 @@ chardet==3.0.4            # via requests
 contextlib2==0.5.5        # via raven
 cryptography==2.4.2       # via pyopenssl, requests
 directory-api-client==13.0.0
-directory-ch-client==1.0.0
-directory-client-core==4.3.0  # via directory-api-client, directory-ch-client, directory-sso-api-client
-directory-components==8.3.1
-directory-constants==14.4.0  # via directory-components
+directory-client-core==4.3.0  # via directory-api-client, directory-sso-api-client
+directory-components==15.1.1
+directory-constants==16.7.2
 directory-healthcheck==1.0.1
 directory-sso-api-client==4.1.1
 directory_validators==5.0.0
-django-admin-ip-restrictor==1.0.0  # via directory-components
 django-cors-headers==2.1.0
 django-environ==0.4.5
 django-formtools==2.0
 django-health-check==3.8.0  # via directory-healthcheck
-django-ipware==2.1.0      # via django-admin-ip-restrictor
 django-redis==4.8.0
 django-revproxy==0.9.15
 django==1.11.20
 django_extensions==1.9.0
-djangorestframework==3.4.7  # via sigauth
+djangorestframework==3.9.1
 gunicorn==19.5.0
 idna==2.6                 # via cryptography, requests
-mohawk==0.3.4             # via directory-components, sigauth
-monotonic==1.2            # via directory-ch-client, directory-client-core
+jsonschema==3.0.1         # via directory-components
+mohawk==0.3.4             # via sigauth
+monotonic==1.2            # via directory-client-core
 olefile==0.44             # via directory-validators, pillow
-phonenumberslite==7.7.4
-pillow==4.2.1
+phonenumberslite==7.7.4   # via directory-validators
+pillow==4.2.1             # via directory-validators
 pycparser==2.19           # via cffi
 pyopenssl==18.0.0         # via requests
+pyrsistent==0.15.2        # via jsonschema
 pytz==2017.2              # via directory-validators, django
 raven==5.19.0
 redis==2.10.6             # via django-redis
 requests[security]==2.21.0
-sigauth==4.0.1            # via directory-ch-client, directory-client-core
-six==1.11.0               # via cryptography, django-extensions, mohawk, pyopenssl, w3lib
+sigauth==4.0.1            # via directory-client-core
+six==1.11.0               # via cryptography, django-extensions, jsonschema, mohawk, pyopenssl, pyrsistent, w3lib
 urllib3==1.24.2
 w3lib==1.19.0             # via directory-client-core
 whitenoise==3.1
