+++ requirements_test.txt
@@ -6,7 +6,7 @@
 #
 asn1crypto==0.24.0        # via cryptography
 atomicwrites==1.2.1       # via pytest
-attrs==18.2.0             # via pytest
+attrs==18.2.0             # via jsonschema, pytest
 beautifulsoup4==4.6.3     # via directory-components
 certifi==2018.11.29       # via requests
 cffi==1.11.5              # via cryptography
@@ -16,37 +16,34 @@ contextlib2==0.5.5        # via raven
 coverage==4.5.2           # via codecov, pytest-cov
 cryptography==2.4.2       # via pyopenssl, requests
 directory-api-client==13.0.0
-directory-ch-client==1.0.0
-directory-client-core==4.3.0  # via directory-api-client, directory-ch-client, directory-sso-api-client
-directory-components==8.3.1
-directory-constants==14.4.0  # via directory-components
+directory-client-core==4.3.0  # via directory-api-client, directory-sso-api-client
+directory-components==15.1.1
+directory-constants==16.7.2
 directory-healthcheck==1.0.1
 directory-sso-api-client==4.1.1
 directory_validators==5.0.0
-django-admin-ip-restrictor==1.0.0  # via directory-components
 django-cors-headers==2.1.0
 django-environ==0.4.5
 django-formtools==2.0
 django-health-check==3.8.0  # via directory-healthcheck
-django-ipware==2.1.0      # via django-admin-ip-restrictor
 django-redis==4.8.0
 django-revproxy==0.9.15
 django==1.11.20
 django_extensions==1.9.0
-djangorestframework==3.9.0  # via sigauth
+djangorestframework==3.9.1
 flake8==3.6.0
 freezegun==0.3.11
 gunicorn==19.5.0
 idna==2.7                 # via cryptography, requests
+jsonschema==3.0.1         # via directory-components
 mccabe==0.6.1             # via flake8
-mohawk==0.3.4             # via directory-components, sigauth
-monotonic==1.5            # via directory-ch-client, directory-client-core
+mohawk==0.3.4             # via sigauth
+monotonic==1.5            # via directory-client-core
 more-itertools==4.3.0     # via pytest
 olefile==0.44             # via directory-validators, pillow
 packaging==18.0           # via pytest-sugar
-pathlib2==2.3.3           # via pytest
-phonenumberslite==7.7.4
-pillow==4.2.1
+phonenumberslite==7.7.4   # via directory-validators
+pillow==4.2.1             # via directory-validators
 pluggy==0.8.0             # via pytest
 py==1.7.0                 # via pytest
 pycodestyle==2.4.0        # via flake8
@@ -54,6 +51,7 @@ pycparser==2.19           # via cffi
 pyflakes==2.0.0           # via flake8
 pyopenssl==18.0.0         # via requests
 pyparsing==2.3.0          # via packaging
+pyrsistent==0.15.2        # via jsonschema
 pytest-cov==2.6.0
 pytest-django==3.4.4
 pytest-sugar==0.9.2
@@ -64,8 +62,8 @@ raven==5.19.0
 redis==2.10.6             # via django-redis
 requests-mock==1.5.2
 requests[security]==2.21.0
-sigauth==4.0.1            # via directory-ch-client, directory-client-core
-six==1.11.0               # via cryptography, django-extensions, freezegun, mohawk, more-itertools, packaging, pathlib2, pyopenssl, pytest, python-dateutil, requests-mock, w3lib
+sigauth==4.0.1            # via directory-client-core
+six==1.11.0               # via cryptography, django-extensions, freezegun, jsonschema, mohawk, more-itertools, packaging, pyopenssl, pyrsistent, pytest, python-dateutil, requests-mock, w3lib
 termcolor==1.1.0          # via pytest-sugar
 urllib3==1.24.2
 w3lib==1.19.0             # via directory-client-core
